@@ -1,11 +1,13 @@
 import uvicorn
 from main import app
+import logging
 
 if __name__ == "__main__":
     uvicorn.run(
-        uvicorn.run(app, host="localhost", port=8000),
-        reload = True,
-        app = "main:app",
-        log_level=logging.INFO,
+        "main:app", 
+        host="localhost", 
+        port=8000, 
+        reload=True,
+        log_level='trace',
         use_colors=True
-    )
+    ),
