@@ -1,5 +1,5 @@
 'use client'
-import React, {useState, useEffect} from "react";
+import React, {useState} from "react";
 import { UserLoginData } from "./types";
 import styles from './LoginForm.module.css'
 import { useRouter } from 'next/navigation';
@@ -18,8 +18,8 @@ const LoginForm = () => {
     const router = useRouter();
     //create react component with useState - initially blank
     const[userLoginData, setUserLoginData] = useState<UserLoginData>({
-        entered_email: '',
-        entered_password:''
+        email: '',
+        password:''
     });
     
     //send POST request with user data for backend to verify
@@ -34,6 +34,7 @@ const LoginForm = () => {
             });
             // go to root on success
             if(response.ok) {
+                alert("login successful - this is just a placeholder to validate implementation works correctly")
                 router.push("/")
             } else {
                 throw new Error(`${response.status}`);
@@ -41,7 +42,7 @@ const LoginForm = () => {
             
             //LOGGING
             const endpointResponse = await response.json();
-            console.log(endpointResponse)
+            alert(endpointResponse)
 
         } catch(error) {
             //TODO: better error handling 
@@ -57,27 +58,34 @@ const LoginForm = () => {
             <input 
                 className={styles.input}
                 type="text"
-                value={userLoginData.entered_email}
-                onChange={(e) => setUserLoginData({...userLoginData, entered_email:e.target.value})}
+                value={userLoginData.email}
+                placeholder="Enter your email address"
+                onChange={(e) => setUserLoginData({...userLoginData, email:e.target.value})}
             />
             <input 
                 className={styles.input}
                 type="password"
                 name="password"
-                value={userLoginData.entered_password}
-                onChange={(e) => setUserLoginData({...userLoginData, entered_password:e.target.value})}
+                placeholder="Enter your password"
+                value={userLoginData.password}
+                onChange={(e) => setUserLoginData({...userLoginData, password:e.target.value})}
             />
             <button className={styles.button} type='submit'>Login</button>
         </form>
         
-            <Link  href="/">
+            <Link href="/">
                 <h1 className={styles.OutgoingLink}>
-                Go Home
+                    Go Home
                 </h1>
             </Link>
             <Link href="/auth/register">
                 <h1 className={styles.OutgoingLink}>
-                Register for an account here
+                    Register for an account here
+                </h1>
+            </Link>
+            <Link href="/auth/check_password_token">
+                <h1 className={styles.OutgoingLink}>
+                    Reset Password
                 </h1>
             </Link>
         </div>
