@@ -1,4 +1,4 @@
 export interface UserLoginData{
-    entered_email:string;
-    entered_password:string;
+    email:string;
+    password:string;
 }
