@@ -1,3 +1,5 @@
 export interface ResetPasswordData{
-    new_password: string
+    new_password: string,
+    confirm_password: string,
+    reset_token: string
 }
