@@ -6,8 +6,7 @@ export default function Home() {
   return (
     <main>
       <h1 className={styles.title}>Root Page</h1>
-      <h1 className= {styles.reg}><Link href = "/auth/register">Go to Registration</Link></h1>
-      <h1 className= {styles.log}><Link href = "/auth/login">Go to Login</Link></h1>
+      <h1 className= {styles.log}><Link href = "/auth/login">Register/Login</Link></h1>
     </main>
   )
 }
